move subcommands to main dir
