+++ cmd/newtestercmd/main.go
@@ -11,9 +11,9 @@ import (
 	"time"
 
 	"github.com/urfave/cli/v3"
-	"gitlab.com/aquachain/aquachain/cmd/aquachain/aquaflags"
 	"gitlab.com/aquachain/aquachain/common/log"
 	"gitlab.com/aquachain/aquachain/internal/debug"
+	"gitlab.com/aquachain/aquachain/subcommands/aquaflags"
 )
 
 var errMainQuit = errors.New("run success")
