+++ subcommands/bugcmd.go
