+++ subcommands/buildinfo/buildinfo.go
