+++ subcommands/mnemoniccmd.go
